refactor: comment_selectorを機能別モジュールに分割
- 53KB/1099行の巨大ファイルを機能別に分割
- base_selector.py: メインのCommentSelectorクラス
- llm_selector.py: LLM関連の選択ロジック
- validation.py: バリデーションロジック
- utils.py: ユーティリティ関数
- 後方互換性を維持（既存のインポートは変更不要）
