+++ src/nodes/comment_selector/__init__.py
@@ -1,0 +1,5 @@
+"""コメント選択モジュール"""
+
+from .base_selector import CommentSelector
+
+__all__ = ['CommentSelector']
